Separate app from server (index)
